backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.5
web3==6.1.0
drf-spectacular[sidecar]==0.26.1
pytest-django==4.5.2